+++ src/components/motion/motion.tsx
@@ -202,7 +202,7 @@ const Motion = ({ parentId, onMotionChanged, onTotalChanged, parentType, current
               startFrom={parentType === "tomograms" ? "middle" : "end"}
               total={motion.total || motion.rawTotal}
               onChange={setPage}
-              displayDefault={page}
+              defaultPage={page}
             />
           </>
         )}