+++ src/components/spa/classification.test.tsx
@@ -2,7 +2,7 @@ import { Classification } from "./classification";
 import { renderWithProviders } from "../../utils/test-utils";
 import { fireEvent, screen, waitFor } from "@testing-library/react";
 
-describe("2D Classification", () => {
+describe("Classification", () => {
   window.URL.createObjectURL = jest.fn();
   it("should display first page as default", async () => {
     renderWithProviders(<Classification autoProcId={1} />);
@@ -37,10 +37,10 @@ describe("2D Classification", () => {
     });
   });
 
-  it("should display message when no classification data is available", async () => {
+  it("should not display row when classification data is not available", async () => {
     renderWithProviders(<Classification autoProcId={2} />);
 
-    await screen.findByText("No Classification Data Found");
+    await waitFor(() => expect(screen.queryByText("2D Classification")).not.toBeInTheDocument());
   });
 
   it("should display visualisation button for 3D", async () => {
@@ -71,5 +71,15 @@ describe("2D Classification", () => {
     fireEvent.click(screen.getByText(/Open 3D Visualisation/i));
 
     await screen.findByText("No Valid Volume File");
+  });
+
+  it("should match selected class to 3D visualisation modal page", async () => {
+    renderWithProviders(<Classification autoProcId={1} type='3d' />);
+
+    fireEvent.click(await screen.findByText(/Open 3D Visualisation/i));
+
+    await waitFor(async () => expect((await screen.findAllByLabelText("Total Pages"))[1]).toHaveTextContent("2"));
+    fireEvent.click((await screen.findAllByLabelText("Next Page"))[1]);
+    await waitFor(() => expect(screen.getByLabelText("Batch Number Value")).toHaveTextContent("355"));
   });
 });