+++ src/components/spa/main.tsx
@@ -1,9 +1,4 @@
-import {
-  AccordionItem,
-  AccordionPanel,
-  Grid,
-  GridItem,
-} from "@chakra-ui/react";
+import { AccordionItem, AccordionPanel, Grid, GridItem } from "@chakra-ui/react";
 import { Motion } from "../motion/motion";
 import { useCallback, useState } from "react";
 import { Classification } from "./classification";
@@ -11,7 +6,6 @@ import { ParticlePicking } from "./particlePicking";
 import { CTF } from "../ctf/ctf";
 import { ProcessingTitle } from "../visualisation/processingTitle";
 import { BaseProcessingJobProps } from "../../schema/interfaces";
-
 
 const SPA = ({ autoProc, procJob, status, active }: BaseProcessingJobProps) => {
   const [total, setTotal] = useState(0);