+++ src/utils/loaders/user.test.tsx
@@ -48,4 +48,15 @@ describe("User Data", () => {
     await getUser();
     expect(global.window.history.replaceState).toBeCalled()
   });
+
+  it("should return null if network request fails", async () => {
+    server.use(
+      rest.get("http://localhost/auth/user", (req, res, ctx) => {
+        return res.networkError("Failed to connect");
+      })
+    );
+
+    const user = await getUser();
+    expect(user).toBe(null)
+  });
 });