+++ CHANGELOG.rst
@@ -3,6 +3,18 @@ Changelog
 ==========
 
 ++++++++++
+v0.8.0 (14/03/2023)
+++++++++++
+
+**Added**
+
+- Adds pagination to 3D visualisation modal
+
+**Changed**
+
+- 3D visualisation modal does not try to render model twice
+
+++++++++++
 v0.7.0 (07/03/2023)
 ++++++++++